Rename symlinkManager.ts to integrationManager.ts
