+++ src/integrations/integrationManager.ts
